+++ CHANGELOG.md
@@ -40,6 +40,9 @@
   - Lunaris do not try to compute anymore POST->DELETE and UPDATE->DELETE offline transactions.
   - Add lazy loading for stores. Use `isLazyLoad = true` in stores to activate it.
   - Change `lunaris.getOne` to use primary key: `lunaris.getOne(store, value, isPrimaryKey)`.
+  - Add `lunaris.websocket.subscribe({String} channel, {Function} handler)`.
+  - Add `lunaris.websocket.unsubscribe({String} channel)`.
+  - Add `vm.socketChannels` to subscribe to websocket channels.
 - Fix:
   - Improve offline app.
   - In offline or offline mode, generated urls were not encoded.